requirements.txt
# Bot Discord
discord.py>=2.3.0

# LLM + Embeddings (Google Gemini)
google-genai>=1.0.0

# Supabase (HTTP client direto - sem dependências pesadas)
httpx>=0.26.0

# Document parsing
PyPDF2>=3.0.0
python-docx>=1.1.0

# Text splitting
langchain-text-splitters>=0.3.0

# Web parsing
beautifulsoup4>=4.12.0

# Structured extraction
langextract>=1.0.0

# Microsoft Teams (Bot Framework)
botbuilder-core>=4.17.0
botbuilder-integration-aiohttp>=4.17.0
aiohttp>=3.9.0

# Utils
python-dotenv>=1.0.0
